ui/requirements.txt
paho-mqtt==1.3.1
respeaker==0.5.0
raven==6.0.0
